Eslint, Material-UI, deployment workflow integration (#1)
* test build

* Styling on homepage

I was testing deployment and found some bug on Material-ui with gatsby.
The style is inconsistent based on how the user naviagate to the page.
i.e. visit directly with URL and visit via clicking Link.
Fixed the issue with more restrictive styling.
Need to do more rigorous testing.

* adding material-ui plugin

* Use Helmet to include font dependency

Main purpose of using helmet is to remove git dependency on the
`public/` folder. And fonts are directly included in the code.

* Adding eslint and fixing style inconsistency

* Adding git hook to lint before commit

* Change `format` npm option, update regex for posts

* Return array instead of fragment in Homepage

* Adapt MentorQ style linting

* Adding material-ui style injection

* More fixes

* Bump to ES2019

* More comments