MenuBar new adaptive threshold
Due to the addition of the Events tab, we need to increase the
breakpoint where the MenuBar changes from desktop to mobile
layout. But built-in breakpoint in M-UI offers a granularity
that is too low. Therefore, we use its value but multiply it
to get to a comfortable breakpoint for our design.

The upside of this approach is that it is highly extensible.
If later we decide to remove or add tabs, we can simply change
this scaling factor.

The downside of this approach is that it makes MenuBar special
in a way that it changes between mobile and desktop layout at
a different screen size compared to other components. But this
does not really hurts user experience since user almost never
changes screen size.