Revamp SVG support
Create a fork of gatsby-plugin-svgr that minifies SVGs, even if they are
not inlined. (They were previously only minified if the ReactComponent
named export is used.)

With the exception of hackdescription.js due to a browser bug, this
commit institutes a new policy for including SVG resources:

- If the file is small, then it should be included statically using the
  ReactComponent named export.
- If the SVG file is relatively large, then it should be included using
  the new SVGImg component, which sets up lazy loading and prevents
  unnecessary reflows/flashes.

Previously, a third option was available, which was including the SVG as
a "data:" URL. However, all such instances should now just be included
statically, and this option is eliminated by our fork of
gatsby-plugin-svgr.

Fixes: #46